# app/Http/OpenapiSchemas/PostUpdateRequest.php
<?php

namespace App\Http\OpenapiSchemas;

/**
 * Class UpdateRequest.
 *
 * @OA\Schema(
 *     title="UpdateRequest model",
 *     description="UpdateRequest model",
 * )
 */
class PostUpdateRequest extends PostCreateRequest
{

}