# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Post API Demo",
        "version": "1.0.0"
    },
    "paths": {
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Category"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login",
                "description": "Login",
                "operationId": "login",
                "requestBody": {
                    "description": "product",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "deprecated": false
            }
        },
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "getPosts",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostIndexResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Posts",
                "description": "Posts",
                "operationId": "createPost",
                "requestBody": {
                    "description": "post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContent"
                                }
                            }
                        }
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Get post",
                "description": "Return post",
                "operationId": "getPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Posts",
                "description": "Posts",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "post",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PostUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableContent"
                                }
                            }
                        }
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Delete post",
                "description": "Delete post",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get list of users",
                "description": "Returns list of users",
                "operationId": "getUsers",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserIndexResponse"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Category": {
                "title": "Category model",
                "description": "Category model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "example": "Technology"
                    },
                    "image_url": {
                        "title": "Image URL",
                        "description": "Image URL",
                        "type": "string",
                        "example": "http://localhost:8000/images/covers/technology.jpg"
                    },
                    "icon_url": {
                        "title": "Icon URL",
                        "description": "Icon URL",
                        "type": "string",
                        "example": "http://localhost:8000/images/icons/technology.png"
                    }
                },
                "type": "object"
            },
            "LinkPagination": {
                "title": "LinkPagination model",
                "description": "LinkPagination model",
                "properties": {
                    "url": {
                        "title": "Url",
                        "description": "Url",
                        "type": "string",
                        "example": "null"
                    },
                    "label": {
                        "title": "Label",
                        "description": "Label",
                        "type": "string",
                        "example": "&laquo; Previous"
                    },
                    "active": {
                        "title": "Active",
                        "description": "Active",
                        "type": "boolean",
                        "example": "false"
                    }
                },
                "type": "object"
            },
            "LoginRequest": {
                "title": "LoginRequest model",
                "description": "LoginRequest model",
                "properties": {
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "example": "correo1@example.com"
                    },
                    "password": {
                        "title": "Password",
                        "description": "Password",
                        "type": "string",
                        "example": "password"
                    }
                },
                "type": "object"
            },
            "LoginResponse": {
                "title": "LoginResponse model",
                "description": "LoginResponse model",
                "properties": {
                    "token": {
                        "title": "Token",
                        "description": "Token",
                        "type": "string",
                        "example": "1|USMA41xVLp4Zwy7XpULYDuZywbZBE7EoC5BhXtMH"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string",
                        "example": "Success"
                    }
                },
                "type": "object"
            },
            "PaginationResponse": {
                "title": "PaginationResponse model",
                "description": "PaginationResponse model",
                "properties": {
                    "current_page": {
                        "title": "current_page",
                        "description": "current_page",
                        "type": "integer",
                        "format": "int64"
                    },
                    "first_page_url": {
                        "title": "First page url.",
                        "description": "First page url.",
                        "type": "string"
                    },
                    "from": {
                        "title": "from",
                        "description": "from",
                        "type": "integer",
                        "format": "int64"
                    },
                    "last_page": {
                        "title": "last_page",
                        "description": "last_page",
                        "type": "integer",
                        "format": "int64"
                    },
                    "last_page_url": {
                        "title": "Last page url.",
                        "description": "Last page url.",
                        "type": "string"
                    },
                    "links": {
                        "title": "links",
                        "description": "links",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/LinkPagination"
                        }
                    },
                    "next_page_url": {
                        "title": "Next page url.",
                        "description": "Next page url.",
                        "type": "string"
                    },
                    "path": {
                        "title": "Path.",
                        "description": "Path.",
                        "type": "string"
                    },
                    "per_page": {
                        "title": "Per page.",
                        "description": "Number of items per page.",
                        "type": "integer",
                        "format": "int64"
                    },
                    "prev_page_url": {
                        "title": "Previus page url.",
                        "description": "Previus page url.",
                        "type": "string"
                    },
                    "to": {
                        "title": "to",
                        "description": "to",
                        "type": "integer",
                        "format": "int64"
                    },
                    "total": {
                        "title": "total",
                        "description": "total",
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "type": "object"
            },
            "Post": {
                "title": "Post model",
                "description": "Post model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "title": {
                        "title": "Title",
                        "description": "Title",
                        "type": "string",
                        "example": "Title of post"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet ..."
                    },
                    "user_id": {
                        "title": "User ID",
                        "description": "User ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "user": {
                        "$ref": "#/components/schemas/User"
                    },
                    "category_id": {
                        "title": "Category ID",
                        "description": "Category ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "category": {
                        "$ref": "#/components/schemas/Category"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "example": "2022-05-20T17:05:32.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "example": "2022-05-20T17:05:32.000000Z"
                    }
                },
                "type": "object"
            },
            "PostCreateRequest": {
                "title": "PostCreateRequest model",
                "description": "PostCreateRequest model",
                "required": [
                    "title",
                    "description",
                    "category_id"
                ],
                "properties": {
                    "title": {
                        "title": "Title",
                        "description": "Title",
                        "type": "string",
                        "example": "Title of post"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Description",
                        "type": "string",
                        "example": "Lorem ipsum dolor sit amet ..."
                    },
                    "category_id": {
                        "title": "Category ID",
                        "description": "Category ID",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "PostIndexResponse": {
                "title": "PostIndexResponse model",
                "description": "PostIndexResponse model",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationResponse"
                    },
                    {
                        "properties": {
                            "data": {
                                "title": "List of posts",
                                "description": "List of posts",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/Post"
                                }
                            }
                        }
                    }
                ]
            },
            "PostUpdateRequest": {
                "title": "UpdateRequest model",
                "description": "UpdateRequest model",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PostCreateRequest"
                    }
                ]
            },
            "UnprocessableContent": {
                "title": "UnprocessableContent model",
                "description": "UnprocessableContent model",
                "properties": {
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "title": "Errors",
                        "description": "Errors",
                        "example": "{}",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "User": {
                "title": "User model",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "ID",
                        "description": "ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "name": {
                        "title": "Name",
                        "description": "Name",
                        "type": "string",
                        "example": "Desmond Adams"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email",
                        "type": "string",
                        "example": "correo1@example.com"
                    },
                    "email_verified_at": {
                        "title": "Email verified at",
                        "description": "Email verified at",
                        "type": "string",
                        "example": "2022-05-20T17:05:32.000000Z"
                    },
                    "created_at": {
                        "title": "Created at",
                        "description": "Created at",
                        "type": "string",
                        "example": "2022-05-20T17:05:32.000000Z"
                    },
                    "updated_at": {
                        "title": "Updated at",
                        "description": "Updated at",
                        "type": "string",
                        "example": "2022-05-20T17:05:32.000000Z"
                    }
                },
                "type": "object"
            },
            "UserIndexResponse": {
                "title": "UserIndexResponse model",
                "description": "UserIndexResponse model",
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/PaginationResponse"
                    },
                    {
                        "properties": {
                            "data": {
                                "title": "List of users",
                                "description": "List of users",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                ]
            }
        },
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}